test(api-headless-cms): skip nested query test in dynamodb only tests
